move exchangeOpenIdUserInfo() into helper.go
